+++ pkg/disttask/framework/meter/meter.go
@@ -47,7 +47,7 @@ func RecordDXFS3GetRequests(store kv.Storage, getReqCnt uint64) {
 		return
 	}
 	userKS := store.GetKeyspace()
-	meteringInstance.Load().Record(userKS, &MeterData{getRequests: uint64(getReqCnt)})
+	meteringInstance.Load().Record(userKS, &MeterData{getRequests: getReqCnt})
 }
 
 // RecordDXFS3PutRequests records the S3 PUT requests for DXF.
@@ -56,25 +56,25 @@ func RecordDXFS3PutRequests(store kv.Storage, putReqCnt uint64) {
 		return
 	}
 	userKS := store.GetKeyspace()
-	meteringInstance.Load().Record(userKS, &MeterData{putRequests: uint64(putReqCnt)})
+	meteringInstance.Load().Record(userKS, &MeterData{putRequests: putReqCnt})
 }
 
 // RecordDXFScanDataTraffic records the scan data traffic for DXF.
-func RecordDXFScanDataTraffic(store kv.Storage, size int64) {
+func RecordDXFScanDataTraffic(store kv.Storage, size uint64) {
 	if !kv.IsUserKS(store) || meteringInstance.Load() == nil {
 		return
 	}
 	userKS := store.GetKeyspace()
-	meteringInstance.Load().Record(userKS, &MeterData{scanDataTraffic: uint64(size)})
+	meteringInstance.Load().Record(userKS, &MeterData{scanDataTraffic: size})
 }
 
 // RecordDXFWriteDataSize records the write data size for DXF.
-func RecordDXFWriteDataSize(store kv.Storage, size int64) {
+func RecordDXFWriteDataSize(store kv.Storage, size uint64) {
 	if !kv.IsUserKS(store) || meteringInstance.Load() == nil {
 		return
 	}
 	userKS := store.GetKeyspace()
-	meteringInstance.Load().Record(userKS, &MeterData{writeDataSize: uint64(size)})
+	meteringInstance.Load().Record(userKS, &MeterData{writeDataSize: size})
 }
 
 // GetMetering gets the metering instance.
@@ -203,10 +203,10 @@ func (m *Meter) flush(ts int64, timeout time.Duration) {
 			"version":           "1",
 			"cluster_id":        keyspace,
 			"source_name":       category,
-			"s3_put_requests":   &common.MeteringValue{Value: uint64(d.putRequests), Unit: "count"},
-			"s3_get_requests":   &common.MeteringValue{Value: uint64(d.getRequests), Unit: "count"},
-			"scan_data_traffic": &common.MeteringValue{Value: uint64(d.scanDataTraffic), Unit: "bytes"},
-			"write_data_size":   &common.MeteringValue{Value: uint64(d.writeDataSize), Unit: "bytes"},
+			"s3_put_requests":   &common.MeteringValue{Value: d.putRequests, Unit: "count"},
+			"s3_get_requests":   &common.MeteringValue{Value: d.getRequests, Unit: "count"},
+			"scan_data_traffic": &common.MeteringValue{Value: d.scanDataTraffic, Unit: "bytes"},
+			"write_data_size":   &common.MeteringValue{Value: d.writeDataSize, Unit: "bytes"},
 		})
 	}
